fix: redis7.0 监控查询calls数值过大转Intger异常

+++ yudao-module-infra/yudao-module-infra-biz/src/main/java/cn/iocoder/yudao/module/infra/convert/redis/RedisConvert.java
@@ -22,7 +22,7 @@ public interface RedisConvert {
         commandStats.forEach((key, value) -> {
             respVO.getCommandStats().add(RedisMonitorRespVO.CommandStat.builder()
                     .command(StrUtil.subAfter((String) key, "cmdstat_", false))
-                    .calls(Integer.valueOf(StrUtil.subBetween((String) value, "calls=", ",")))
+                    .calls(Long.valueOf(StrUtil.subBetween((String) value, "calls=", ",")))
                     .usec(Long.valueOf(StrUtil.subBetween((String) value, "usec=", ",")))
                     .build());
         });
